requirements.txt
pandas
faker
psycopg2-binary
openpyxl